Trabalho Final/aeroporto/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.6
dnspython==2.2.1
pymongo==4.1.1
python-dotenv==0.19.2
sqlparse==0.4.2
